# backend_services/gateway/docs/swagger.yaml
basePath: /
definitions:
  github_com_eternal-orbit-labs_gateway_internal_domain.App:
    properties:
      category:
        type: string
      created_at:
        type: string
      description:
        type: string
      icon:
        type: string
      id:
        type: string
      name:
        type: string
      slug:
        type: string
      sort_order:
        type: integer
      status:
        type: string
      tagline:
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_eternal-orbit-labs_gateway_internal_domain.PersonalAPIKey:
    properties:
      created_at:
        type: string
      id:
        type: string
      label:
        type: string
      last_used_at:
        type: string
      prefix:
        type: string
      user_id:
        type: string
    type: object
...
  github_com_eternal-orbit-labs_gateway_internal_domain.User:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      phone:
        type: string
      updated_at:
        type: string
    type: object
...
  internal_handler.AppGetResponse:
    properties:
      app:
        $ref: '#/definitions/github_com_eternal-orbit-labs_gateway_internal_domain.App'
    type: object
  internal_handler.AppListResponse:
    properties:
      apps:
        items:
          $ref: '#/definitions/github_com_eternal-orbit-labs_gateway_internal_domain.App'
        type: array
      next_cursor:
        example: ""
        type: string
    type: object
...
  internal_handler.ChangePasswordRequest:
    properties:
      current_password:
        example: oldP@ss1
        type: string
      new_password:
        example: newSecureP@ss1
        minLength: 8
        type: string
    required:
    - new_password
    type: object
  internal_handler.CreateAPIKeyRequest:
    properties:
      label:
        example: my-key
        type: string
    required:
    - label
    type: object
  internal_handler.CreateAPIKeyResponse:
    properties:
      key:
        $ref: '#/definitions/github_com_eternal-orbit-labs_gateway_internal_domain.PersonalAPIKey'
      raw_value:
        example: eol_k1_abc123...
        type: string
    type: object
...
  internal_handler.ErrorResponse:
    properties:
      error:
        example: error message
        type: string
    type: object
...
  internal_handler.ListAPIKeysResponse:
    properties:
      keys:
        items:
          $ref: '#/definitions/github_com_eternal-orbit-labs_gateway_internal_domain.PersonalAPIKey'
        type: array
    type: object
  internal_handler.LoginRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: secureP@ss1
        type: string
    required:
    - email
    - password
    type: object
...
  internal_handler.MessageResponse:
    properties:
      message:
        example: operation completed
        type: string
    type: object
...
  internal_handler.SignupRequest:
    properties:
      email:
        example: user@example.com
        type: string
      first_name:
        example: John
        type: string
      last_name:
        example: Doe
        type: string
      password:
        example: secureP@ss1
        minLength: 8
        type: string
    required:
    - email
    - first_name
    - password
    type: object
  internal_handler.SuccessResponse:
    properties:
      success:
        example: true
        type: boolean
    type: object
...
  internal_handler.UpdateMeRequest:
    properties:
      first_name:
        example: Jane
        type: string
      last_name:
        example: Smith
        type: string
    type: object
...
  internal_handler.UserResponse:
    properties:
      user:
        $ref: '#/definitions/github_com_eternal-orbit-labs_gateway_internal_domain.User'
    type: object
host: localhost:8080
info:
  contact: {}
  description: API gateway for Eternal Orbit Labs — authentication, organizations,
    teams, and app management.
  title: EOL Gateway API
  version: "1.0"
paths:
  /api/apps:
    get:
      description: Returns a paginated list of apps. Supports cursor-based pagination
        and search.
      parameters:
      - description: Search query
        in: query
        name: q
        type: string
      - description: Pagination cursor
        in: query
        name: cursor
        type: string
      - default: 20
        description: Results per page (max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.AppListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - CookieAuth: []
      - BearerAPIKey: []
      summary: List apps
      tags:
      - Apps
  /api/apps/{slug}:
    get:
      description: Returns a single app by its URL slug.
      parameters:
      - description: App slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.AppGetResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - CookieAuth: []
      - BearerAPIKey: []
      summary: Get app by slug
      tags:
      - Apps
...
  /api/auth/apikeys:
    get:
      description: Returns all personal API keys for the authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.ListAPIKeysResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - CookieAuth: []
      - BearerAPIKey: []
      summary: List API keys
      tags:
      - API Keys
    post:
      consumes:
      - application/json
      description: Generate a new personal API key. The raw value is only returned
        once.
      parameters:
      - description: API key label
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.CreateAPIKeyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handler.CreateAPIKeyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - CookieAuth: []
      - BearerAPIKey: []
      summary: Create API key
      tags:
      - API Keys
  /api/auth/apikeys/{id}:
    delete:
      description: Permanently delete a personal API key by ID.
      parameters:
      - description: API Key ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - CookieAuth: []
      - BearerAPIKey: []
      summary: Delete API key
      tags:
      - API Keys
  /api/auth/change-password:
    post:
      consumes:
      - application/json
      description: Change the authenticated user's password. Requires current password
        (unless the account was created via OAuth).
      parameters:
      - description: Password change payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - CookieAuth: []
      - BearerAPIKey: []
      summary: Change password
      tags:
      - Auth
...
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate with email and password. Sets httpOnly auth cookies
        on success.
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Log in
      tags:
      - Auth
...
  /api/auth/me:
    get:
      description: Returns the authenticated user's profile.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - CookieAuth: []
      - BearerAPIKey: []
      summary: Get current user
      tags:
      - Auth
    patch:
      consumes:
      - application/json
      description: Update the authenticated user's first name and/or last name.
      parameters:
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.UpdateMeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - CookieAuth: []
      - BearerAPIKey: []
      summary: Update profile
      tags:
      - Auth
  /api/auth/refresh:
    post:
      description: Exchange the refresh token cookie for a new access token. Rotates
        the refresh token.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Refresh access token
      tags:
      - Auth
  /api/auth/signup:
    post:
      consumes:
      - application/json
      description: Register a new user with email and password. Sets httpOnly auth
        cookies on success.
      parameters:
      - description: Signup credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.SignupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handler.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Create a new account
      tags:
      - Auth
schemes:
- http
- https
securityDefinitions:
  BearerAPIKey:
    description: Personal API key — send as "Bearer eol_k1_..."
    in: header
    name: Authorization
    type: apiKey
  CookieAuth:
    description: JWT access token set automatically on login/signup.
    in: cookie
    name: eol_access
    type: apiKey
swagger: "2.0"